+++ app/services/statistics/fetcher_service.rb
@@ -34,6 +34,7 @@ class Statistics::FetcherService
                       duration: p.find_statistic_type.key,
                       subscription_id: sub.id,
                       invoice_item_id: i.id,
+                      coupon: i.invoice.coupon&.code,
                       ca: ca }.merge(user_info(profile)))
       end
       result
@@ -59,7 +60,8 @@ class Statistics::FetcherService
                    slot_dates: r.slots.map(&:start_at).map(&:to_date),
                    nb_hours: (r.slots.map(&:duration).map(&:to_i).reduce(:+) / 3600.0).to_f,
                    ca: calcul_ca(r.original_invoice),
-                   reservation_context_id: r.reservation_context_id
+                   reservation_context_id: r.reservation_context_id,
+                   coupon: r.original_invoice.coupon&.code
                   }.merge(user_info(profile))
         yield result
       end
@@ -85,7 +87,8 @@ class Statistics::FetcherService
                    slot_dates: r.slots.map(&:start_at).map(&:to_date),
                    nb_hours: (r.slots.map(&:duration).map(&:to_i).reduce(:+) / 3600.0).to_f,
                    ca: calcul_ca(r.original_invoice),
-                   reservation_context_id: r.reservation_context_id
+                   reservation_context_id: r.reservation_context_id,
+                   coupon: r.original_invoice.coupon&.code
                   }.merge(user_info(profile))
         yield result
       end
@@ -112,7 +115,8 @@ class Statistics::FetcherService
                    training_date: slot.start_at.to_date,
                    nb_hours: difference_in_hours(slot.start_at, slot.end_at),
                    ca: calcul_ca(r.original_invoice),
-                   reservation_context_id: r.reservation_context_id
+                   reservation_context_id: r.reservation_context_id,
+                   coupon: r.original_invoice&.coupon&.code
                  }.merge(user_info(profile))
         yield result
       end
@@ -128,6 +132,7 @@ class Statistics::FetcherService
         .eager_load(:slots, :slots_reservations, :invoice_items, statistic_profile: [:group])
         .find_each do |r|
         next unless r.reservable
+        next unless r.original_invoice
 
         profile = r.statistic_profile
         slot = r.slots.first
@@ -141,6 +146,7 @@ class Statistics::FetcherService
                    age_range: (r.reservable.age_range_id ? r.reservable.age_range.name : ''),
                    nb_places: r.total_booked_seats,
                    nb_hours: difference_in_hours(slot.start_at, slot.end_at),
+                   coupon: r.original_invoice.coupon&.code,
                    ca: calcul_ca(r.original_invoice) }.merge(user_info(profile))
         yield result
       end
@@ -215,7 +221,7 @@ class Statistics::FetcherService
            .where('order_activities.created_at >= :start_date AND order_activities.created_at <= :end_date', options)
            .group('orders.id')
            .find_each do |o|
-        result = { date: o.created_at.to_date, ca: calcul_ca(o.invoice) }
+        result = { date: o.created_at.to_date, ca: calcul_ca(o.invoice), coupon: o.invoice.coupon&.code }
                  .merge(user_info(o.statistic_profile))
                  .merge(store_order_info(o))
         yield result