+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fix a bug: unable to cancel a payment schedule
 - adds reservation context feature (for machine, training, space)
+- adds coupon in statistic export (for subscription, machine, training, space, event, order)
 - [TODO DEPLOY] `rails db:seed`
 - [TODO DEPLOY] `rails fablab:es:build_stats`
 - [TODO DEPLOY] `rails fablab:maintenance:regenerate_statistics[2014,1]`